DragonLib 3 update

+++ common/src/main/java/de/mrjulsen/crn/data/schedule/condition/TrainSeparationCondition.java
@@ -148,7 +148,7 @@ public class TrainSeparationCondition extends ScheduledDelay implements IDelayed
 		return worldTime + totalWaitTicks();
 	}
 
-    public static DLTime toTime(int ticks) {
+    public static DLTime toTime(long ticks) {
 		return DLTime.fromTicks(ticks, new ConfiguredTimeSystem());
     }
